+++ GUI/include/heatequationsolver.h
@@ -17,31 +17,31 @@ public:
 
     HeatEquationSolver(const InputData &data);
 
-    QVector<float> temperatures() const;
+    QVector<double> temperatures() const;
     QVector<CoordDiff> xoffsets() const;
     QVector<CoordDiff> yoffsets() const;
     QVector2D origin() const;
-    QVector<float> initialConditions() const;
-    QMap<int, float> boundaryConditions_1() const;
+    QVector<double> initialConditions() const;
+    QMap<int, double> boundaryConditions_1() const;
     QMap<QPair<int, int>, ConvectionData> boundaryConditions_3() const;
 
     struct ConvectionData {
-        float m_emissionCoef;
-        float m_ambientTemperature;
+        double m_emissionCoef;
+        double m_ambientTemperature;
     };
 
 private:
-    QVector<float> m_temperatures;
+    QVector<double> m_temperatures;
     QVector<CoordDiff> m_xoffsets;
     QVector<CoordDiff> m_yoffsets;
-    QVector<float> m_initialConditions;
-    QMap<int, float> m_boundaryConditions_1;
+    QVector<double> m_initialConditions;
+    QMap<int, double> m_boundaryConditions_1;
     QMap<QPair<int, int>, ConvectionData> m_boundaryConditions_3;
 
     QVector2D m_origin;
 };
 
-inline QVector<float> HeatEquationSolver::temperatures() const {
+inline QVector<double> HeatEquationSolver::temperatures() const {
     return m_temperatures;
 }
 
@@ -56,12 +56,12 @@ inline QVector2D HeatEquationSolver::origin() const {
     return m_origin;
 }
 
-inline QVector<float> HeatEquationSolver::initialConditions() const
+inline QVector<double> HeatEquationSolver::initialConditions() const
 {
     return m_initialConditions;
 }
 
-inline QMap<int, float> HeatEquationSolver::boundaryConditions_1() const
+inline QMap<int, double> HeatEquationSolver::boundaryConditions_1() const
 {
     return m_boundaryConditions_1;
 }
