+++ HeatEquationSolver/src/heatequationsolver.cpp
@@ -21,13 +21,13 @@ QVector<QVector<T>> stdToQtMatrix(const std::vector<std::vector<T>>& m)
     return res;
 }
 
-std::map<int, float> getBoundCond(
+std::map<int, double> getBoundCond(
     Coord x0, Coord y0,
-    const std::vector<float>& hx,
-    const std::vector<float>& hy,
+    const std::vector<double>& hx,
+    const std::vector<double>& hy,
     MathFunc phi)
 {
-    std::map<int, float> cond1;
+    std::map<int, double> cond1;
     Coord x = x0;
     Coord y = y0;
 
@@ -62,10 +62,10 @@ HeatEquationSolver::HeatEquationSolver(const InputData& data) {
         origin, data.k_x, data.k_y, data.f);
 
     const auto cond1 = getBoundCond(data.x0, data.y0, data.hx, data.hy, data.phi);
-    m_boundaryConditions_1 = QMap<int, float>(cond1);
+    m_boundaryConditions_1 = QMap<int, double>(cond1);
     countBorderRules(k_v.first, k_v.second, cond1);
-    QVector<QVector<float>> K = stdToQtMatrix(k_v.first);
-    QVector<float> V = QVector<float>::fromStdVector(k_v.second);
+    QVector<QVector<double>> K = stdToQtMatrix(k_v.first);
+    QVector<double> V = QVector<double>::fromStdVector(k_v.second);
 
     m_origin = QVector2D(origin[0], origin[1]);
 
@@ -75,7 +75,7 @@ HeatEquationSolver::HeatEquationSolver(const InputData& data) {
     m_valid = solveSystem(K, V, m_temperatures);
 }
 
-QVector<float> HeatEquationSolver::temperatures() const
+QVector<double> HeatEquationSolver::temperatures() const
 {
     return m_temperatures;
 }
@@ -95,12 +95,12 @@ QVector2D HeatEquationSolver::origin() const
     return m_origin;
 }
 
-QVector<float> HeatEquationSolver::initialConditions() const
+QVector<double> HeatEquationSolver::initialConditions() const
 {
     return m_initialConditions;
 }
 
-QMap<int, float> HeatEquationSolver::boundaryConditions_1() const
+QMap<int, double> HeatEquationSolver::boundaryConditions_1() const
 {
     return m_boundaryConditions_1;
 }
